requirement.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.3
fastapi==0.72.0
grpcio==1.43.0
grpcio-tools==1.43.0
h11==0.13.0
idna==3.3
protobuf==3.19.4
pydantic==1.8.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.0
